@@ -2,6 +2,7 @@
     "name": "vue3-course",
     "version": "0.1.0",
     "private": true,
+    "type": "module",
     "scripts": {
         "serve": "vue-cli-service serve",
         "build": "vue-cli-service build",
@@ -17,24 +18,35 @@
         "volar-service-vetur": "^0.0.14",
         "vue": "^3.2.36",
         "vue-router": "^4.0.3",
-        "vuex": "^4.0.2"
+        "vuex": "^4.0.2",
+        "vuex-persistedstate": "^4.1.0"
     },
     "devDependencies": {
         "@babel/core": "^7.23.2",
         "@babel/eslint-parser": "^7.12.16",
         "@babel/preset-env": "^7.23.2",
         "@types/uuid": "^9.0.6",
+        "@typescript-eslint/eslint-plugin": "^6.12.0",
+        "@typescript-eslint/parser": "^6.12.0",
         "@vue/cli-plugin-babel": "~5.0.0",
         "@vue/cli-plugin-eslint": "~5.0.0",
         "@vue/cli-plugin-router": "~5.0.0",
+        "@vue/cli-plugin-typescript": "~5.0.0",
         "@vue/cli-plugin-vuex": "~5.0.0",
         "@vue/cli-service": "^5.0.8",
         "@vue/compiler-sfc": "^3.3.7",
-        "eslint": "^8.51.0",
+        "@vue/eslint-config-typescript": "^9.1.0",
+        "css-loader": "^6.8.1",
+        "eslint": "^8.54.0",
         "eslint-config-prettier": "^8.3.0",
         "eslint-plugin-prettier": "^5.0.1",
         "eslint-plugin-vue": "^9.17.0",
-        "prettier": "^3.0.3"
+        "prettier": "^3.0.3",
+        "sass": "^1.69.5",
+        "sass-loader": "^13.3.2",
+        "style-loader": "^3.3.3",
+        "typescript": "^5.3.2",
+        "vite": "^5.0.2"
     },
     "eslintConfig": {
         "root": true,
@@ -44,10 +56,11 @@
         "extends": [
             "plugin:vue/vue3-essential",
             "eslint:recommended",
-            "plugin:prettier/recommended"
+            "plugin:prettier/recommended",
+            "@vue/typescript"
         ],
         "parserOptions": {
-            "parser": "@babel/eslint-parser"
+            "parser": "@typescript-eslint/parser"
         },
         "rules": {}
     },
