@@ -1,16 +1,14 @@
 <template>
-    <div class="container convertation-history__container">
-        <h1 class="convertation-history__header">История конвертаций:</h1>
-        <ButtonForConvert
-            class="button button__cleanup-history"
-            @click="clearHistory"
-            >Очистить историю</ButtonForConvert
+    <div class="convertation-history container">
+        <h1 class="convertation-history header">История конвертаций:</h1>
+        <ButtonComponent button-style="cleanup-history" @click="clearHistory"
+            >Очистить историю</ButtonComponent
         >
-        <transition-group name="convertation-history__list"
+        <transition-group name="convertation-history list"
             ><div
                 v-for="item in getCurrenciesHistory"
                 :key="item"
-                class="container list-item"
+                class="convertation-history list-item"
             >
                 {{ item }}
             </div>
@@ -18,72 +16,79 @@
     </div>
 </template>
 
-<script>
-import { mapGetters } from 'vuex'
-export default {
-    name: 'ConvertationHistory',
+<script setup lang="ts">
+import { onMounted, computed } from 'vue';
+import { useStore } from 'vuex';
 
-    computed: { ...mapGetters(['getCurrenciesHistory']) },
-    mounted() {},
-    methods: {
-        clearHistory() {
-            localStorage.setItem('convertListItemsArray', '[]')
-        },
-    },
-}
+const store = useStore();
+
+const getCurrenciesHistory = computed(
+    () => store.getters['convert/getCurrenciesHistory'],
+);
+
+const deleteCurrenciesHistory = () => {
+    store.dispatch('convert/deleteCurrenciesHistory');
+};
+
+const fetchCurrenciesHistory = () => {
+    store.dispatch('convert/fetchCurrenciesHistory');
+};
+const cleanCurrenciesHistory = (): void => {
+    store.commit('convert/cleanCurrenciesHistory');
+};
+
+onMounted(() => {
+    fetchCurrenciesHistory();
+});
+
+const clearHistory = (): void => {
+    cleanCurrenciesHistory();
+    deleteCurrenciesHistory();
+};
 </script>
 
-<style scoped>
-.container.convertation-history__container {
+<style scoped lang="scss">
+.convertation-history.container {
     top: 80px;
     right: 0;
     padding: 15px;
-    width: 350px;
+    width: fit-content;
     height: calc(100vh - 156px);
-    border: 4px solid #12c0b2;
-    overflow-y: scroll;
-}
-.convertation-history__header {
-    margin-bottom: 15px;
-    font-size: 24px;
-    text-align: center;
-}
+    border-left: 4px solid #12c0b2;
+    overflow-y: auto;
+    transition: transform 0.3s ease-in-out;
+    transform: translateX(0);
 
-.button.button__cleanup-history {
-    margin-bottom: 15px;
-    width: 150px;
-    height: 30px;
-    font-size: 12px;
-    padding: 5px;
-}
-.convertation-history__list {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    list-style: none;
-}
-.container.list-item {
-    margin-bottom: 15px;
-    padding: 15px;
-    width: 100%;
-}
-.list-item {
-    text-align: center;
-    padding: 5px;
-    width: 100%;
-    border: 2px solid #12c0b2;
-    border-radius: 5px;
-}
-.list-item:not(:last-of-type) {
-    margin-bottom: 10px;
-}
-.convertation-history__list-enter-active,
-.convertation-history__list-leave-active {
-    transition: all 1s;
-}
-.convertation-history__list-enter-from,
-.convertation-history__list-leave-to {
-    opacity: 0;
-    transform: translateX(30px);
+    .header {
+        margin-bottom: 15px;
+        font-size: 24px;
+    }
+
+    .list {
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        list-style: none;
+    }
+    .list-item {
+        text-align: center;
+        padding: 15px;
+        width: 100%;
+        border: 2px solid #12c0b2;
+        border-radius: 5px;
+
+        &:not(:last-of-type) {
+            margin-bottom: 10px;
+        }
+    }
+    .list-enter-active,
+    .list-leave-active {
+        transition: all 1s;
+    }
+    .list-enter-from,
+    .list-leave-to {
+        opacity: 0;
+        transform: translateX(30px);
+    }
 }
 </style>
