@@ -1,43 +1,55 @@
 <template>
-    <div class="input__container">
+    <div class="currency-input__container">
         <input
             v-focus
-            :value="modelValue"
+            :value="props.modelValue"
             class="input"
             type="number"
             placeholder="Введите сумму"
             @input="updateInput"
         />
-        <small-button @click="clearInput"> Очистить </small-button>
+        <ButtonComponent
+            button-style="currency-input_cleanup"
+            @click="clearInput"
+        >
+            Очистить
+        </ButtonComponent>
     </div>
 </template>
 
-<script>
-export default {
-    name: 'CurrencyInput',
-    props: {
-        modelValue: {
-            type: [String, Number],
-            default: '',
-            required: true,
-        },
-    },
-    emits: ['update:modelValue', 'clearInput'],
+<script setup lang="ts">
+import { PropType } from 'vue';
 
-    methods: {
-        updateInput(event) {
-            this.$emit('update:modelValue', event.target.value)
-        },
-        clearInput() {
-            this.$emit('update:modelValue', '')
-            localStorage.removeItem('amount')
-        },
+const props = defineProps({
+    modelValue: {
+        type: [String, Number] as PropType<string | number>,
+        required: true,
     },
-}
+});
+const emit = defineEmits(['update:modelValue']);
+
+const updateInput = (input: Event): void => {
+    const inputValue = (input.target as HTMLInputElement).value;
+    emit('update:modelValue', inputValue);
+};
+
+const clearInput = (): void => {
+    emit('update:modelValue', '');
+};
 </script>
 
-<style scoped>
+<style scoped lang="scss">
+.currency-input__container {
+    display: flex;
+    flex-direction: row;
+    gap: 0;
+}
+
 .input {
+    @extend .currency-input__input;
+}
+
+.currency-input__input {
     position: relative;
     width: 100%;
     border: 2px solid #18aa66;
@@ -45,21 +57,15 @@ export default {
     border-radius: 10px 0 0 10px;
     font-size: 16px;
     outline: none;
-}
 
-input::-webkit-outer-spin-button,
-input::-webkit-inner-spin-button {
-    -webkit-appearance: none;
-    margin: 0;
-}
+    &::-webkit-outer-spin-button,
+    &::-webkit-inner-spin-button {
+        -webkit-appearance: none;
+        margin: 0;
+    }
 
-input[type='number'] {
-    appearance: textfield; /* Firefox */
-}
-
-.input__container {
-    display: flex;
-    flex-direction: row;
-    gap: 0;
+    &[type='number'] {
+        appearance: textfield; /* Firefox */
+    }
 }
 </style>
