@@ -1,4 +1,38 @@
-const { defineConfig } = require('@vue/cli-service')
-module.exports = defineConfig({
+import { defineConfig } from '@vue/cli-service';
+import path from 'path';
+
+const generalVarsFiles = ['/assets/scss/_variables.scss'];
+
+export default defineConfig({
     transpileDependencies: true,
-})
+    css: {
+        loaderOptions: {
+            scss: {
+                additionalData: async (content, loaderContext) => {
+                    const { resourcePath, rootContext } = loaderContext;
+
+                    const relativePath = path.relative(
+                        rootContext,
+                        resourcePath,
+                    );
+                    const varsFiles = [...generalVarsFiles];
+
+                    const varsFilesImport = varsFiles.map(
+                        (item) => `@import "@${item}";`,
+                    );
+
+                    const skipVarFiles = varsFilesImport.some((item) => {
+                        const fileName = item.match(/\/(_.+\.scss)";$/)[1];
+                        return relativePath.includes(fileName);
+                    });
+
+                    if (skipVarFiles) {
+                        return;
+                    }
+
+                    return '\n' + varsFilesImport.join('\n') + content;
+                },
+            },
+        },
+    },
+});
