@@ -3,11 +3,12 @@ module.exports = {
     env: {
         node: true,
         browser: true,
+        es2021: true,
     },
     extends: [
-        'plugin:vue/vue3-recommended',
         'eslint:recommended',
-        'plugin:prettier/recommended',
+        'plugin:@typescript-eslint/recommended',
+        'prettier',
     ],
     rules: {
         'prettier/prettier': [
@@ -23,8 +24,9 @@ module.exports = {
     globals: {
         $nuxt: true,
     },
+    parser: 'vue-eslint-parser',
     parserOptions: {
-        parser: '@babel/eslint-parser',
-        requireConfigFile: false,
+        parser: '@typescript-eslint/parser',
     },
+    plugins: ['@typescript-eslint', 'prettier'],
 };
