@@ -1,44 +1,52 @@
 <template>
     <div class="app">
-        <component :is="currentLayout">
+        <component :is="currentComponent">
             <router-view />
         </component>
     </div>
 </template>
 
-<script>
-import currencyLayout from '@/layouts/currencyLayout.vue'
-import defaultLayout from '@/layouts/defaultLayout.vue'
-import { mapMutations } from 'vuex'
-export default {
-    name: 'App',
-    components: {
-        currencyLayout,
-        defaultLayout,
-    },
-    computed: {
-        currentLayout() {
-            return this.$route.meta.layout
-        },
-    },
+<script setup lang="ts">
+import { onMounted } from 'vue';
+import { useStore } from 'vuex';
+import { useRoute } from 'vue-router';
+import currencyLayout from '@/layouts/currencyLayout.vue';
+import defaultLayout from '@/layouts/defaultLayout.vue';
+import { computed } from 'vue';
 
-    mounted() {
-        this.initExchangeHistory()
-    },
-    methods: {
-        ...mapMutations(['setCurrenciesHistory']),
-        initExchangeHistory() {
-            const exchangeHistory = JSON.parse(
-                localStorage.getItem('convertListItemsArray'),
-            )
-            const exchangeHistoryArray = exchangeHistory
-            this.setCurrenciesHistory(exchangeHistoryArray)
-        },
-    },
-}
+const components: { [key: string]: any } = {
+    currencyLayout,
+    defaultLayout,
+};
+
+const store = useStore();
+const route = useRoute();
+
+const currentComponent = computed(() => {
+    return components[route.meta.layout as string];
+});
+
+const addConvertListItemToHistoryArray = (convertListItem: string): void => {
+    store.commit('convert/addConvertListItemToHistoryArray', convertListItem);
+};
+
+const fetchCurrenciesHistory = () => {
+    store.dispatch('convert/fetchCurrenciesHistory');
+};
+
+const initExchangeHistory = (): void => {
+    const exchangeHistory: any = fetchCurrenciesHistory();
+
+    const exchangeHistoryArray: string = exchangeHistory;
+    addConvertListItemToHistoryArray(exchangeHistoryArray);
+};
+
+onMounted(() => {
+    initExchangeHistory();
+});
 </script>
 
-<style>
+<style lang="scss">
 * {
     margin: 0;
     padding: 0;
