@@ -1,24 +1,20 @@
 <template>
-    <footer class="footer">
+    <footer class="footer-component">
         <div className="copyright">
             <p>
                 © 2023 All rights reserved <br />
-                <a href="https://kreker2112.github.io/goit_js_basic_course/">
-                    kreker2112.github.io
+                <a href="https://github.com/kreker2112/-urrency-converter.git">
+                    kreker2112.github
                 </a>
             </p>
         </div>
     </footer>
 </template>
 
-<script>
-export default {
-    name: 'FooterComponent',
-}
-</script>
+<script setup lang="ts"></script>
 
-<style scoped>
-.footer {
+<style scoped lang="scss">
+.footer-component {
     position: absolute;
     bottom: 0;
     width: 100%;
@@ -26,13 +22,22 @@ export default {
     padding: 20px 0 20px;
     background-color: #12c0b2;
     color: #fff;
-}
-.copyright {
-    font-size: 16px;
-    text-align: center;
-}
 
-.copyright p + p {
-    margin-top: 0.5em;
+    .copyright {
+        gap: 5px;
+        font-size: 16px;
+        text-align: center;
+
+        p {
+            display: flex;
+            flex-direction: column;
+            gap: 5px;
+        }
+
+        a {
+            color: #273baa;
+            text-decoration: underline;
+        }
+    }
 }
 </style>
