@@ -4,25 +4,31 @@
             Личный кабинет предпринимателя
         </div>
         <div class="navbar__btns">
-            <nav-button @click="$router.push('/converter')"
-                >Конвертер валют</nav-button
+            <ButtonComponent
+                button-style="nav-button"
+                v-for="button in navButtons"
+                :key="button.id"
+                @click="$router.push(button.link)"
+                >{{ button.name }}</ButtonComponent
             >
-
-            <nav-button @click="$router.push('/#')">Поступления</nav-button>
-            <nav-button @click="$router.push('/#')"
-                >Налоговые обязанности</nav-button
-            >
-            <nav-button @click="$router.push('/#')">Уплата налогов</nav-button>
-            <nav-button @click="$router.push('/#')">О сайте</nav-button>
         </div>
     </div>
 </template>
 
-<script>
-export default {}
+<script setup lang="ts">
+import { v4 as uuidv4 } from 'uuid';
+import { ref } from 'vue';
+
+const navButtons = ref([
+    { id: uuidv4(), name: 'Конвертер валют', link: '/converter' },
+    { id: uuidv4(), name: 'Поступления', link: '/funds' },
+    { id: uuidv4(), name: 'Налоговые обязанности', link: '/#' },
+    { id: uuidv4(), name: 'Уплата налогов', link: '/#' },
+    { id: uuidv4(), name: 'О сайте', link: '/#' },
+]);
 </script>
 
-<style scoped>
+<style scoped lang="scss">
 .navbar {
     position: absolute;
     width: 100%;
@@ -35,23 +41,22 @@ export default {}
     gap: 15px;
     padding: 0 15px;
     cursor: pointer;
-}
-.navbar_title {
-    font-size: 2rem;
-    font-family: 'Lucida Sans', 'Lucida Sans Regular', 'Lucida Grande',
-        'Lucida Sans Unicode', Geneva, Verdana, sans-serif;
-    color: #fff;
-    padding: 15px;
-    height: 80px;
-}
-
-.navbar__btns {
-    position: absolute;
-    right: 0;
-    display: flex;
-    align-items: center;
-    justify-content: flex-end;
-    gap: 15px;
-    padding-right: 20px;
+    .navbar_title {
+        font-size: 2rem;
+        font-family: 'Lucida Sans', 'Lucida Sans Regular', 'Lucida Grande',
+            'Lucida Sans Unicode', Geneva, Verdana, sans-serif;
+        color: #fff;
+        padding: 15px;
+        height: 80px;
+    }
+    .navbar__btns {
+        position: absolute;
+        right: 0;
+        display: flex;
+        align-items: center;
+        justify-content: flex-end;
+        gap: 15px;
+        padding-right: 20px;
+    }
 }
 </style>
