@@ -2,18 +2,18 @@
     <Navbar />
     <div class="currency-layout">
         <slot></slot>
-        <ConvertationHistory />
+        <ConvertationHistory class="convertation-history__component" />
     </div>
     <FooterComponent />
 </template>
 
-<script setup>
-import ConvertationHistory from '@/components/ConvertationHistory.vue'
-import Navbar from '@/components/NavBar.vue'
-import FooterComponent from '@/components/FooterComponent.vue'
+<script setup lang="ts">
+import ConvertationHistory from '@/components/ConvertationHistory.vue';
+import Navbar from '@/components/NavBar.vue';
+import FooterComponent from '@/components/FooterComponent.vue';
 </script>
 
-<style scoped>
+<style scoped lang="scss">
 .currency-layout {
     padding-top: 80px;
     display: flex;
