@@ -1,20 +1,19 @@
 <template>
-    <div class="input-form__container">
-        <InputForm />
+    <div class="converter-page">
+        <div class="input-form__container">
+            <InputForm />
+        </div>
     </div>
 </template>
 
-<script>
-import InputForm from '@/components/InputForm.vue'
-export default {
-    name: 'ConverterPage',
-    components: {
-        InputForm,
-    },
-}
+<script setup lang="ts">
+import InputForm from '@/components/InputForm.vue';
 </script>
 
-<style scoped>
+<style scoped lang="scss">
+.converter-page {
+    margin: 0 auto;
+}
 .input-form__container {
     margin: 0 auto;
 }
